ultima entrega - visualizar

+++ mercado/models.py
@@ -9,6 +9,6 @@ class Supermercado(models.Model):
     produto = models.CharField(max_length=50, default=None)
     categoria = models.CharField(max_length=50, default=None)
     quantidade = models.IntegerField(null=True, default=None)
-    preço = models.IntegerField(null=True, default=None)
-
+    #preço = models.IntegerField(null=True, default=None)
+    preço = models.DecimalField(null=True, max_digits=5, decimal_places=2, default=None)
